src/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
awscli-local==0.20
boto3==1.28.31
botocore==1.31.31
bson==0.5.10
certifi==2023.7.22
cffi==1.15.1
click==8.1.7
fastapi==0.101.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
localstack-client==2.0
packaging==23.1
pluggy==1.2.0
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
pytest==7.4.0
python-dateutil==2.8.2
python-decouple==3.8
PyYAML==6.0
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
